+++ etna/transforms/math/differencing.py
@@ -8,10 +8,15 @@ from typing import cast
 import numpy as np
 import pandas as pd
 
+from etna import SETTINGS
 from etna.datasets import TSDataset
 from etna.transforms.base import ReversibleTransform
 from etna.transforms.utils import check_new_segments
 from etna.transforms.utils import match_target_quantiles
+
+if SETTINGS.auto_required:
+    from optuna.distributions import BaseDistribution
+    from optuna.distributions import IntUniformDistribution
 
 
 class _SingleDifferencingTransform(ReversibleTransform):
@@ -473,3 +478,17 @@ class DifferencingTransform(ReversibleTransform):
         for transform in self._differencing_transforms[::-1]:
             result_df = transform._inverse_transform(result_df)
         return result_df
+
+    def params_to_tune(self) -> Dict[str, "BaseDistribution"]:
+        """Get default grid for tuning hyperparameters.
+
+        This grid tunes only ``order`` parameter. Other parameters are expected to be set by the user.
+
+        Returns
+        -------
+        :
+            Grid to tune.
+        """
+        return {
+            "order": IntUniformDistribution(low=1, high=2),
+        }